--- packages/nest/src/index.ts
@@ -1,8 +1,0 @@
-export * from './implement'
-export {
-  /**
-   * @alias Implement
-   */
-  Implement as Impl,
-} from './implement'
-export * from './module'
