+++ packages/nest/src/openapi/implement.test-d.ts
@@ -2,7 +2,7 @@ import { oc } from '@orpc/contract'
 import { implement, os } from '@orpc/server'
 import { describe, it } from 'vitest'
 import { z } from 'zod'
-import { Implement } from './implement'
+import { Implement } from './index'
 
 export const inputSchema = z.object({ input: z.number().transform(n => `${n}`) })
