+++ packages/nest/src/openapi/module.test.ts
@@ -7,8 +7,7 @@ import { openapi } from '@orpc/openapi'
 import { implement } from '@orpc/server'
 import supertest from 'supertest'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
-import { Implement } from './implement'
-import { ORPCModule } from './module'
+import { Implement, ORPCModule } from './index'
 
 beforeEach(() => {
   vi.clearAllMocks()
